whiteboard_backend/SAVEDrequirements.txt
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cu126
--extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cu126/

accelerate==1.12.0

aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
bce-python-sdk==0.9.56
beautifulsoup4==4.14.2
hf-xet 
bitsandbytes==0.49.0
cachetools==6.2.4
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
Django==5.2.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
duckduckgo_search==8.1.1
einops==0.8.1
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
greenlet==3.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.4
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
lxml==6.0.2
MarkupSafe==3.0.3
modelscope==1.33.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
nltk==3.9.2
num2words==0.5.14
numpy==2.2.6

opencv-contrib-python==4.10.0.84

opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==25.0

paddleocr==3.3.2
paddlex==3.3.12
paddlepaddle-gpu==3.2.0

torch==2.6.0+cu126
torchvision==0.21.0+cu126
torchaudio==2.6.0+cu126

pandas==2.3.3
pillow==12.0.0
playwright==1.57.0
prettytable==3.17.0
primp==0.15.0
propcache==0.4.1
protobuf==3.20.2
psutil==7.2.0
psycopg==3.2.6
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
PyJWT==2.9.0
pypdfium2==5.2.0
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pyvips-binary==8.18.0
PyYAML==6.0.2
regex==2025.10.23
requests==2.32.5
requests-file==3.0.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
selenium==4.38.0
sentence-transformers==5.2.0
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
sqlparse==0.5.5
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.12.20
tim==0.5
timm==1.0.22
tldextract==5.3.1
tokenizers==0.22.1
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers@c7e5b749a6392ea2f42fea983af41f825b0bc78d
trio==0.32.0
trio-websocket==0.12.2
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.14
websocket-client==1.9.0
wsproto==1.3.2
yarl==1.22.0
